Update Provider.php
YouTube's channel list service is no longer consistent. It now returns a response without an items array or with an empty items array. A completely missing items array means that the google/gmail account was selected as an option. An empty items array is returned for suspended, banned or removed channels. In times past YouTube would automatically make a channel for your google/gmail account. This is no longer the case so we cannot assume that items is an array or that items will have any indexes.